refactor: update topology and gyard rendering
