+++ src/WebClient/app/components/graveyard/index.ts
@@ -1,0 +1,1 @@
+export { default as Graveyard } from './Graveyard';
