+++ src/WebClient/app/components/layout/Socials.tsx
@@ -1,6 +1,8 @@
-import { Button, Image, makeStyles, tokens } from '@fluentui/react-components';
-import useDarkMode from '@microsoftgraveyard/hooks/useDarkMode';
 import { JSX } from 'react';
+
+import { Button, Image, makeStyles, tokens } from '@fluentui/react-components';
+
+import useDarkMode from '@microsoftgraveyard/hooks/useDarkMode';
 
 interface ISocial {
   href: string;
