+++ src/WebClient/app/hooks/useDarkMode.ts
@@ -1,5 +1,6 @@
-import { DarkModeContext } from '@microsoftgraveyard/providers/DarkMode';
 import { useContext } from 'react';
+
+import { DarkModeContext } from '@microsoftgraveyard/providers/DarkMode';
 
 const useDarkMode = () => useContext(DarkModeContext);
