+++ src/WebClient/app/providers/index.ts
@@ -1,0 +1,2 @@
+export { default as DarkModeProvider } from './DarkMode';
+export { default as ThemeProvider } from './Theme';
