+++ src/WebClient/app/types/Corpse.ts
@@ -1,0 +1,10 @@
+interface Corpse {
+  name: string;
+  qualifier: string | null;
+  birthDate: Date | null;
+  deathDate: Date;
+  description: string;
+  link: string;
+}
+
+export type { Corpse };
