+++ src/WebClient/app/types/index.ts
@@ -1,0 +1,1 @@
+export type { Corpse } from './Corpse';
