+++ src/WebClient/package.json
@@ -5,36 +5,37 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "tsc --build && vite build",
+    "build": "tsc --noEmit && vite build",
     "lint": "eslint --fix .",
     "lint:check": "eslint .",
     "preview": "vite preview"
   },
   "dependencies": {
-    "@fluentui/react-components": "^9.58.1",
-    "@fluentui/react-icons": "^2.0.274",
+    "@fluentui/react-components": "^9.60.0",
+    "@fluentui/react-icons": "^2.0.279",
     "@griffel/react": "^1.5.29",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-responsive": "^10.0.0"
+    "react-responsive": "^10.0.1"
   },
   "devDependencies": {
-    "@eslint/js": "^9.20.0",
+    "@eslint/js": "^9.21.0",
     "@griffel/tag-processor": "^1.0.12",
     "@griffel/vite-plugin": "^0.1.12",
-    "@types/node": "^22.13.4",
+    "@trivago/prettier-plugin-sort-imports": "^5.2.2",
+    "@types/node": "^22.13.9",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "@vitejs/plugin-react": "^4.3.4",
-    "eslint": "^9.17.0",
+    "eslint": "^9.21.0",
     "eslint-plugin-prettier": "^5.2.3",
     "eslint-plugin-react": "^7.37.4",
-    "eslint-plugin-react-hooks": "^5.0.0",
-    "eslint-plugin-react-refresh": "^0.4.16",
-    "globals": "^15.15.0",
-    "prettier": "^3.4.2",
-    "typescript": "~5.7.3",
-    "typescript-eslint": "^8.18.2",
-    "vite": "^6.0.5"
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
+    "prettier": "^3.5.3",
+    "typescript": "^5.8.2",
+    "typescript-eslint": "^8.25.0",
+    "vite": "^6.2.1"
   }
 }
