+++ src/WebClient/vite.config.ts
@@ -1,7 +1,7 @@
-import { defineConfig } from 'vite';
-import plugin from '@vitejs/plugin-react';
 import griffel from '@griffel/vite-plugin';
+import plugin from '@vitejs/plugin-react';
 import { resolve } from 'path';
+import { defineConfig } from 'vite';
 
 const config = defineConfig(({ command }) => ({
   plugins: [plugin(), command === 'build' && griffel()],
